requirements.txt
click
colorlog
booby-ng
asyncssh
python-libnmap